feat(pyr): pyr alias points to ptpython
I tend to check things in the Python repl as I program, and a version
that supports autocompletion and multi line functionality is a better
fit for this habit than the python interpreter.

There are a few options for more complex python repls.
I have aliased pyr (python-repl) to whatever choice I have made.