fix: <Win>-t and <Win>-y create splits
This commit changes <Win>-t from creating a horizontal split to creating
a vertical split, and <Win>-y from creating a vertical split to creating
a horizontal split.

The change is somewhat arbitrary, but makes the chord consistent over
all my keyboards, and is selected for being the closest viable key
combination to the i3 defaults.